AsyncAPI 3.x: the individual gets its own test suite
Review against docs/for_developers.md: a suite is named after the class it
exercises, and the four tests of AsyncApiIndividual were sitting in
AsyncApiActionBuilderTest. Moved, unchanged, to AsyncApiIndividualTest.
